saas-agent-team/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saas-agent-team"
version = "1.0.0"
description = "Multi-agent AI team for autonomous SaaS development with cost-optimized LLM routing"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "SaaS Agent Team" }]

dependencies = [
    # LLM & Agent frameworks
    "crewai>=0.80",
    "langgraph>=0.2",
    "litellm>=1.40",
    "anthropic>=0.40",
    "openai>=1.50",

    # API framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",

    # Memory & storage
    "weaviate-client>=4.0",
    "redis>=5.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "llama-index>=0.11",

    # Task queue
    "celery[redis]>=5.3",

    # Tools
    "e2b>=1.0",
    "PyGithub>=2.0",
    "firecrawl-py>=1.0",
    "tavily-python>=0.3",

    # Observability
    "langsmith>=0.1.70",
    "prometheus-client>=0.20",
    "structlog>=24.1",

    # HTTP client
    "httpx>=0.27",

    # WebSocket
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "pytest-cov>=4.1",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.9",
    "black>=24.2",
    "pre-commit>=3.6",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "A", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests (batch API, etc.)",
]

[tool.black]
line-length = 100
target-version = ["py311"]
